Remove old command codes and cleanup plugin.yml.
