+++ src/main/java/com/onarandombox/MultiverseNetherPortals/commands/NetherPortalCommand.java
@@ -1,22 +1,13 @@
 package com.onarandombox.MultiverseNetherPortals.commands;
 
-//import com.onarandombox.MultiverseNetherPortals.MultiverseNetherPortals;
-//import com.onarandombox.commandhandler.Command;
-//import org.bukkit.command.CommandSender;
-//
-//import java.util.List;
-//
-//public abstract class NetherPortalCommand extends Command {
-//
-//    protected MultiverseNetherPortals plugin;
-//
-//    public NetherPortalCommand(MultiverseNetherPortals plugin) {
-//        super(plugin);
-//        this.plugin = plugin;
-//
-//    }
-//
-//    @Override
-//    public abstract void runCommand(CommandSender arg0, List<String> arg1);
-//
-//}
+import com.onarandombox.MultiverseNetherPortals.MultiverseNetherPortals;
+import com.onarandombox.acf.BaseCommand;
+
+public abstract class NetherPortalCommand extends BaseCommand {
+
+    protected final MultiverseNetherPortals plugin;
+
+    protected NetherPortalCommand(MultiverseNetherPortals plugin) {
+        this.plugin = plugin;
+    }
+}
